requirements.txt
python-docx>=0.5